Licence Header & linter fixes
